I have changed the folloiwng.
240B-4th.tex is changed to 240B-5th.tex
Added UCI-1.jpg, but I don't think this is needed in the Pretext because it is for cover sheet.

I muted or changed labels so there are no mult-labels.

I DIDN'T change the missing reference because I use the "elegantbook" \LaTeX package. The new version of the "elegantbook" disabled all my references but once Pretext is running, should be back.

There are a lot of warning messages because of the "elegantbook" package I don't know how to correct. That is one of the reasons I turn to Pretext. Hopefuly the warming message would go away.